added class and documentation for controller
